Finished... circle is missing
Done and working

+++ rectangle.h
@@ -7,9 +7,9 @@
 using std::string;
 
 class Rectangle: public Shape{
-	size_t _width, _height;
+	double _width, _height;
 public:
-	Rectangle(size_t width, size_t height);
+	Rectangle(double width, double height);
 	double Area()const;
 	double Perimeter()const;
 	string ToString()const;
